Boat uses the Body loader class to load in a body made in the 2d editor Component Class moved
