+++ core/assets/objects/ships/playerShip.json
@@ -4,78 +4,94 @@
       "name": "playership",
       "imagePath": null,
       "origin": {
-        "x": 0.4749999940395355,
-        "y": 0.875
+        "x": 0.5,
+        "y": 1.075
       },
       "polygons": [
         [
           {
-            "x": 0.7750000357627869,
-            "y": 0.550000011920929
+            "x": 0.9000000357627869,
+            "y": 1.5
           },
           {
             "x": 1,
-            "y": 0.625
+            "y": 1.600000023841858
           },
           {
-            "x": 0.8500000238418579,
-            "y": 1.4250000715255737
+            "x": 1,
+            "y": 1.8000000715255737
           },
           {
-            "x": 0.7250000238418579,
-            "y": 1.5750000476837158
+            "x": 0.925000011920929,
+            "y": 1.9500000476837158
           },
           {
-            "x": 0.4749999940395355,
-            "y": 1.649999976158142
+            "x": 0.75,
+            "y": 2.1000001430511475
+          },
+          {
+            "x": 0.5,
+            "y": 2.1500000953674316
           }
         ],
         [
           {
-            "x": 0.4749999940395355,
-            "y": 1.649999976158142
+            "x": 0.5,
+            "y": 2.1500000953674316
           },
           {
             "x": 0.25,
-            "y": 1.5750000476837158
+            "y": 2.1000001430511475
           },
           {
-            "x": 0.125,
-            "y": 1.4250000715255737
+            "x": 0.10000000149011612,
+            "y": 1.9500000476837158
           },
           {
             "x": 0,
-            "y": 0.625
+            "y": 1.8000000715255737
           },
           {
-            "x": 0.20000000298023224,
-            "y": 0.550000011920929
+            "x": 0,
+            "y": 1.600000023841858
           },
           {
-            "x": 0.7750000357627869,
-            "y": 0.550000011920929
+            "x": 0.10000000149011612,
+            "y": 1.5
+          },
+          {
+            "x": 0.9000000357627869,
+            "y": 1.5
           }
         ],
         [
           {
-            "x": 0.20000000298023224,
-            "y": 0.550000011920929
+            "x": 0.10000000149011612,
+            "y": 1.5
           },
           {
-            "x": 0.2750000059604645,
-            "y": 0.2750000059604645
+            "x": 0.10000000149011612,
+            "y": 0.75
           },
           {
-            "x": 0.4749999940395355,
+            "x": 0.22500000894069672,
+            "y": 0.32500001788139343
+          },
+          {
+            "x": 0.5,
             "y": 0
           },
           {
-            "x": 0.699999988079071,
-            "y": 0.2750000059604645
+            "x": 0.7750000357627869,
+            "y": 0.32500001788139343
           },
           {
-            "x": 0.7750000357627869,
-            "y": 0.550000011920929
+            "x": 0.9000000357627869,
+            "y": 0.75
+          },
+          {
+            "x": 0.9000000357627869,
+            "y": 1.5
           }
         ]
       ],
@@ -85,52 +101,68 @@
           "type": "POLYGON",
           "vertices": [
             {
-              "x": 0.4749999940395355,
+              "x": 0.5,
               "y": 0
             },
             {
-              "x": 0.2750000059604645,
-              "y": 0.2750000059604645
+              "x": 0.22500000894069672,
+              "y": 0.32500001788139343
             },
             {
-              "x": 0.20000000298023224,
-              "y": 0.550000011920929
+              "x": 0.10000000149011612,
+              "y": 0.75
+            },
+            {
+              "x": 0.10000000149011612,
+              "y": 1.5
             },
             {
               "x": 0,
-              "y": 0.625
+              "y": 1.600000023841858
             },
             {
-              "x": 0.125,
-              "y": 1.4250000715255737
+              "x": 0,
+              "y": 1.8000000715255737
+            },
+            {
+              "x": 0.10000000149011612,
+              "y": 1.9500000476837158
             },
             {
               "x": 0.25,
-              "y": 1.5750000476837158
+              "y": 2.1000001430511475
             },
             {
-              "x": 0.4749999940395355,
-              "y": 1.649999976158142
+              "x": 0.5,
+              "y": 2.1500000953674316
             },
             {
-              "x": 0.7250000238418579,
-              "y": 1.5750000476837158
+              "x": 0.75,
+              "y": 2.1000001430511475
             },
             {
-              "x": 0.8500000238418579,
-              "y": 1.4250000715255737
+              "x": 0.925000011920929,
+              "y": 1.9500000476837158
             },
             {
               "x": 1,
-              "y": 0.625
+              "y": 1.8000000715255737
+            },
+            {
+              "x": 1,
+              "y": 1.600000023841858
+            },
+            {
+              "x": 0.9000000357627869,
+              "y": 1.5
+            },
+            {
+              "x": 0.9000000357627869,
+              "y": 0.75
             },
             {
               "x": 0.7750000357627869,
-              "y": 0.550000011920929
-            },
-            {
-              "x": 0.699999988079071,
-              "y": 0.2750000059604645
+              "y": 0.32500001788139343
             }
           ]
         }
