+++ core/src/com/hfentonfearn/entitysystems/InputSystem.java
@@ -12,10 +12,10 @@ import com.badlogic.gdx.InputProcessor;
 import com.hfentonfearn.GameManager;
 import com.hfentonfearn.components.PlayerComponent;
 import com.hfentonfearn.components.VelocityComponent;
-import com.hfentonfearn.ecs.Components;
 import com.hfentonfearn.inputs.Keybinds;
 import com.hfentonfearn.ui.PauseDialog;
 import com.hfentonfearn.utils.AssetLoader;
+import com.hfentonfearn.utils.Components;
 
 import static com.hfentonfearn.utils.Constants.ACCELERATION_DRIVE;
 import static com.hfentonfearn.utils.Constants.ACCELERATION_TURN;
